+++ src/math/epsilon.cpp
@@ -2,7 +2,7 @@
 
 namespace civilarium::math::epsilon {
 
-bool EqualsZero(float x)
+bool EqualsZero(double x)
 {
     return (-kZeroTolerance < x) && (x < kZeroTolerance);
 }
